Unifies brisbane fixtures
